src/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
sqlalchemy==2.0.23
alembic==1.12.1
pymysql==1.1.0
cryptography>=42.0.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pydantic[email]==2.5.0
weasyprint==60.2
qrcode[pil]==7.4.2
python-dotenv==1.0.0
Pillow==10.1.0
jinja2==3.1.2
aiofiles==23.2.1
httpx==0.25.2
pandas==2.1.4 